requirements.txt
Django==5.0.7
django-browser-reload==1.13.0
django-environ==0.11.2
django_components==0.84
django-active-link==0.2.2
django-cotton==0.9.37
django-currentuser==0.8.0
django-widget-tweaks==1.5.0
django-htmx==1.19.0
django-recaptcha==4.0.0
Pillow
django-tailwind==3.8.0
django-autoslug==1.9.9
django-ckeditor==6.7.1
django-import-export==4.1.1
django-filter==24.3
